Initiate the append file.
